pyproject.toml
[project]
name = "cant-be-bothered-ai"
version = "0.1.0"
description = "AI-powered meeting transcription and minutes generation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faster-whisper>=1.2.1",
    "nvidia-cudnn-cu12>=9.14.0.64",
    "rich>=14.2.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.47.0",
]

[project.scripts]
transcribe = "cant_be_bothered.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cant_be_bothered"]
